Unrevivable and Uncloneable Split
