+++ Content.Server/Traits/Assorted/UncloneableSystem.cs
@@ -1,0 +1,21 @@
+using Content.Shared.Cloning.Events;
+using Content.Shared.Traits.Assorted;
+
+namespace Content.Server.Traits.Assorted;
+
+public sealed class UncloneableSystem : EntitySystem
+{
+/// Harmony - New Uncloneable Trait
+    public override void Initialize()
+    {
+        base.Initialize();
+
+       SubscribeLocalEvent<UncloneableComponent, CloningAttemptEvent>(OnCloningAttempt);
+    }
+
+    private void OnCloningAttempt(Entity<UncloneableComponent> ent, ref CloningAttemptEvent args)
+    {
+        if (!ent.Comp.Cloneable)
+            args.Cancelled = true;
+    }
+}
