+++ Content.Server/Traits/Assorted/UnrevivableSystem.cs
@@ -1,20 +1,21 @@
-using Content.Shared.Cloning.Events;
+/// using Content.Shared.Cloning.Events;
 using Content.Shared.Traits.Assorted;
 
 namespace Content.Server.Traits.Assorted;
 
 public sealed class UnrevivableSystem : EntitySystem
 {
-    public override void Initialize()
-    {
-        base.Initialize();
+/// Harmony - Moved to a seperate trait
+///    public override void Initialize()
+///    {
+///        base.Initialize();
+///
+///        SubscribeLocalEvent<UnrevivableComponent, CloningAttemptEvent>(OnCloningAttempt);
+///    }
 
-        SubscribeLocalEvent<UnrevivableComponent, CloningAttemptEvent>(OnCloningAttempt);
-    }
-
-    private void OnCloningAttempt(Entity<UnrevivableComponent> ent, ref CloningAttemptEvent args)
-    {
-        if (!ent.Comp.Cloneable)
-            args.Cancelled = true;
-    }
+///    private void OnCloningAttempt(Entity<UnrevivableComponent> ent, ref CloningAttemptEvent args)
+///    {
+///        if (!ent.Comp.Cloneable)
+///            args.Cancelled = true;
+///    }
 }
