+++ Content.Shared/Traits/Assorted/UnrevivableComponent.cs
@@ -14,11 +14,12 @@ public sealed partial class UnrevivableComponent : Component
     [DataField, AutoNetworkedField]
     public bool Analyzable = true;
 
+    /// Harmony - moved to a seperate trait
     /// <summary>
     /// Can this player be cloned using a cloning pod?
     /// </summary>
-    [DataField, AutoNetworkedField]
-    public bool Cloneable = false;
+    /// [DataField, AutoNetworkedField]
+    /// public bool Cloneable = false;
 
     /// <summary>
     /// The loc string used to provide a reason for being unrevivable
