+++ Resources/Locale/en-US/_Harmony/traits/traits.ftl
@@ -51,3 +51,6 @@ trait-hypophonia-desc = You can only whisper.
 
 trait-illeism-name = Illeism
 trait-illeism-desc = You seem to only be able to refer to yourself by name.
+
+trait-uncloneable-name = Uncloneable
+trait-uncloneable-desc = You are unable to be cloned
